in bootstrap, pointing to tljh branch of jupyhterhub repo, not my fork anymore
removed miniforge components. update installer to point to arm64 mambaforge installer when arm architecture is detected. Added sha256 string for arm installer for mambaforge

sudo for installer line

updated bootstrap to point at my fork

removed sudo trying to figure out what is wrong with install

switched the install repo back to jupyterhub because test on arm and x86 worked so this is ready for pr